Remove API v1 endpoints
