--- app/views/__init__.py
@@ -1,1 +1,0 @@
-"""Copyright 2015 Rafal Kowalski"""
