+++ requirements/common.txt
@@ -1,24 +1,20 @@
 git+https://github.com/eliben/pycparser@release_v2.14
 Flask==0.11.1
-Flask-WTF==0.12
 Flask-Script==2.0.5
 Flask-SQLAlchemy==2.1
 Flask-Migrate==1.8.0
 Flask-Admin==1.4.2
-Flask-Autodoc==0.1.2
 Flask-Login==0.3.2
 Flask-Scrypt==0.1.3.6
-Flask-Cors==2.1.2
 flask-restplus==0.9.2
 Flask-JWT==0.3.2
 Flask-SocketIO==2.5
 eventlet==0.19.0
+pytz
 requests-oauthlib==0.6.1
 boto
 icalendar==3.10
 requests[security]==2.10.0
-WTForms-Components==0.10.0
-WTForms==2.1
 colour==0.1.2
 psycopg2==2.6.2
 xlrd==1.0.0
@@ -29,7 +25,6 @@ itsdangerous==0.24
 humanize==0.5.1
 geoip2
 Flask-HTMLmin
-pytz
 SQLAlchemy-Continuum
 diff-match-patch
 arrow
@@ -58,5 +53,4 @@ python-geoip
 python-geoip-geolite2
 python-pentabarf-xml==0.18
 marrow.mailer
-gsutil
 defusedxml
