+++ requirements/dev.txt
@@ -1,13 +1,10 @@
 -r common.txt
 
 nose==1.3.7
-
 pylint==1.5.5
 pep8==1.7.0
-
 # these prevent SSL warnings in urllib3
 pyOpenSSL==16.0.0
 ndg-httpsclient==0.4.0
 pyasn1==0.1.9
-
 gunicorn==19.6.0
