+++ requirements/prod.txt
@@ -1,4 +1,2 @@
 -r common.txt
-
 gunicorn==19.6.0
-Fabric==1.11.1
